[Bayan] Set up Ban and Rating authorization

+++ RatingService/src/main/java/CS4337/Project/RatingController.java
@@ -3,6 +3,7 @@ package CS4337.Project;
 import java.util.List;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,10 +11,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/ratings")
 public class RatingController {
   private final RatingRepository ratingRepository;
+  private final RatingService ratingService;
 
   @Autowired
-  public RatingController(RatingRepository ratingRepository) {
+  public RatingController(RatingRepository ratingRepository, RatingService ratingService) {
     this.ratingRepository = ratingRepository;
+    this.ratingService = ratingService;
   }
 
   @PostMapping("/add")
@@ -45,6 +48,12 @@ public class RatingController {
   @PutMapping("/update/{id}")
   public ResponseEntity<Map<String, String>> updateRating(
       @PathVariable int id, @RequestBody Map<String, Object> payload) {
+    int ratingOwnerId = ratingRepository.getUserIdByRatingId(id);
+    if (!ratingService.isUserAdmin() && !ratingService.isUserRatingOwner(ratingOwnerId)) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+          .body(Map.of("error", "You do not have permission to edit this rating"));
+    }
+
     try {
       String message = (String) payload.get("message");
       Integer rating = (Integer) payload.get("rating");
@@ -65,6 +74,12 @@ public class RatingController {
 
   @DeleteMapping("/delete/{id}")
   public ResponseEntity<Map<String, String>> deleteRating(@PathVariable int id) {
+    int ratingOwnerId = ratingRepository.getUserIdByRatingId(id);
+    if (!ratingService.isUserAdmin() && !ratingService.isUserRatingOwner(ratingOwnerId)) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+          .body(Map.of("error", "You do not have permission to delete this rating"));
+    }
+
     int result = ratingRepository.deleteRating(id);
     return ResponseEntity.ok(
         Map.of(
